add constructor inputs
